Added condition to check if image has been
If the received data contains key "img" then image is downloaded from the link of firebase storage and the image is then stored in local storage with name as a random number . The path of image is used in the creation of notification object which is then sent to the send notification function. If image is not sent in notification then null is stored. In the two function to send notification if image path is not null the image is sent in notification . Finally the notification object is stored in the database