Add feature to receive image along with notification (#39)
* Add local image file uri to notification Model

when image is sent it is downloaded it is saved to local storage and its address is stored in the notification model in database.

* Update dbhandler

added changes to create a new column in database named COLUMN_IMG_URI and add values to it in functions.

* Add image view 

image view is added in single notification which will display the image if it was sent along with notification.

* update NotificationAdapter

imageview is populated with the image located at the path specified in notification object.

* Update DBHandlerTest 

notification constructor has changed , therefore null has been added as parameter image uri in the constructor

* Added condition to check if image has been

If the received data contains key "img" then image is downloaded from the link of firebase storage and the image is then stored in local storage with name as a random number . The path of image is used in the creation of notification object which is then sent to the send notification function. If image is not sent in notification then null is stored. In the two function to send notification if image path is not null the image is sent in notification . Finally the notification object is stored in the database

* Check if menu item is not null before setting it to enable or disable